+++ src/basic_mpc/identification/__init__.py
@@ -1,7 +1,13 @@
 """Identification grey-box (PEM + Kalman)."""
 
+from basic_mpc.identification.compare import run_compare_r1c1_r2c2
 from basic_mpc.identification.metrics import multi_horizon_scores
 from basic_mpc.identification.pem import fit_r1c1
 from basic_mpc.identification.run import run_identify_r1c1
 
-__all__ = ["fit_r1c1", "multi_horizon_scores", "run_identify_r1c1"]
+__all__ = [
+    "fit_r1c1",
+    "multi_horizon_scores",
+    "run_compare_r1c1_r2c2",
+    "run_identify_r1c1",
+]